requirements.txt
Django
django-debug-toolbar
flake8-django
flake8-isort
Pillow
django-crispy-forms
crispy-bootstrap4
django-modeltranslation
djangorestframework
docutils
drf-yasg
psycopg2-binary
environs
pytils
django-import-export

